pgdatahubnew/pgdh_etl/requirements.txt
sqlalchemy[postgresql_psycopg]>=2.0.0
psycopg>=3.1.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
